Fix #126924: instable fire simulation by default
The cfl_condition default was changed in 9a0a75eff0e43dd009b2c with the intention
of making liquids behave better. However, that degraded the simulations of fire.

Now the property will be set to its old value if the quick smoke operator is used.
Can't think of a great way always use the old value for smoke simulations for now.