Fix #127755: 4.3 regression: Some rigged meshes do not load properly
Revert part of 9530852347275105b8d371e0d12ecf5539e74fb6 as that did not
take into account that the `max` property may actually be smaller than
the `min` property.

I've also taken the liberty to document this fact in some comments.

The fix for the crash when `min == max` is still in place.