Fix #127253: Compositor crop gizmo can produce invalid crop configurations
After 4bf08198a7532160b3e30742231f51256f7412a3 it it no longer necessary to flip gizmo bounds if they are not valid.

Note: it is still possible to create an invalid crop and "losing" the gizmo if it is dragged and dropped outside the compositor canvas. This will be addressed in a different patch.

Pull Request: https://projects.blender.org/blender/blender/pulls/127656